Common argument of procedure functions

Common argument of procedure functions

Arguments

prDataFile

A data frame object of clinical procedure data with at least 3 columns: ID, ICD, and Date. As for date column, the data format should be YYYY/MM/DD or YYYY-MM-DD.

idColName

Column name of ID column in prDataFile. Data type of this argumant should be string without quotation marks.

icdColName

Column name of ICD column in prDataFile. Data type of this argumant should be string without quotation marks.

dateColName

Column name of date column in prDataFile (with date in YYYY/MM/DD or YYYY-MM-DD format). Data type of this argumant should be string without quotation marks.

icdVerColName

(Optional) Column name of ICD-9/10 version recorded in prDataFile. Data format in this column should be numeric 9L or 10L.

icd10usingDate

The date that ICD-10 was started to be used in prDataFile dataset. The data format should be YYYY/MM/DD or YYYY-MM-DD. Necessary if icdVerColName is null.

isDescription

Binary. If true, category description of classification methods will be used in the group column. If false, category name will be used. By default, it is set to be True (standard category description).

Details

Common argument of procedure functions