Data integration for condition era calculation

Conditions era is used to integrate distributed data of clinical records into a single progression record

getConditionEra(
  dxDataFile,
  idColName,
  icdColName,
  dateColName,
  icdVerColName = NULL,
  icd10usingDate = NULL,
  groupDataType = ccs,
  customGroupingTable,
  isDescription = TRUE,
  gapDate = 30,
  selectedCaseFile = NULL
)

Arguments

dxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of dxDataFile.

icdColName
A column for ICD of dxDataFile

dateColName
A column for Date of dxDataFile

icdVerColName
A column for ICD-9/10 version. The format should be numeric 9L or 10L.

icd10usingDate
ICD-10 using date

groupDataType
Four Stratified methods can chosen: CCS ('ccs'), CCS levels ('ccslvl1', 'ccslvl2', 'ccslvl3', 'ccslvl4'), PheWAS ('PheWAS'), comorbidities ('ahrq','charlson', 'elix'), precise or fuzzy customized method ('customGrepIcdGroup', 'customIcdGroup'). Change it to any of the other possible variables, default it set to "ccs".

customGroupingTable
Define the grouping categories. `icdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories. "ICD" defines a list of disease-related ICD codes. `icdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories. "grepICD" defines disease-related ICD code character strings containing a regular expression.

isDescription
Category or description of standard classification methods for ICD diagnostic codes. By default, it set to True (standard description).

gapDate
Length of condition gap, By default, it set to 30 days "30".

selectedCaseFile
Table for selectedCases. Default is 'NULL'

Value

A new data.table based on classifying dxDataFile and calculated condition era by groupDataType for each patient.

Details

Calculate condition era by grouped categories of each patient. Conditions era is used to integrate distributed data of clinical records into a single progression record when the interval of admission data is smaller than the length of condition gap, and these admission data are considered same condition era.

See also

Examples

# sample file for example

head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

# Select case with "Diseases of the urinary system" by level 2 of CCS classification

selectedCaseFile <- selectCases(sampleDxFile, ID, ICD, Date,
                                icdVerColName = NULL,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl2,
                                caseCondition = "Diseases of the urinary system",
                                caseCount = 1)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

# Condition era calculation with case selection

Era1 <- getConditionEra(sampleDxFile, ID, ICD, Date, icdVerColName = NULL, "2015-10-01",
                        groupDataType = CCSlvl3,
                        selectedCaseFile = selectedCaseFile)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

# Define the grouping categories

grepTable <- data.frame(Group = "Chronic kidney disease",
                        grepIcd = "^58|^N18",
                        stringsAsFactors = FALSE)

# Condition era calculation with grouping custom method of code standardization

Era2 <- getConditionEra(sampleDxFile, ID, ICD, Date, icdVerColName = NULL, "2015-10-01",
                        groupDataType = customGrepIcdGroup,
                        customGroupingTable = grepTable)

#> Warning: There is no match diagnostic code with the grepTable

#> Error in -"ICD": invalid argument to unary operator

head(Era1)

#>    ID        CCS_LVL_3_LABEL era firstCaseDate endCaseDate count period
#> 1: A0 Chronic kidney disease   1    2009-07-25  2009-07-25     1 0 days
#> 2: A0 Chronic kidney disease   2    2010-03-28  2010-03-28     1 0 days
#> 3: A0 Chronic kidney disease   3    2010-10-29  2010-10-29     1 0 days
#> 4: A0 Chronic kidney disease   4    2012-04-05  2012-04-05     1 0 days
#> 5: A0 Chronic kidney disease   5    2013-12-20  2013-12-20     1 0 days
#> 6: A1 Chronic kidney disease   1    2006-11-29  2006-11-29     1 0 days
#>    selectedCase
#> 1:     Selected
#> 2:     Selected
#> 3:     Selected
#> 4:     Selected
#> 5:     Selected
#> 6:     Selected

head(Era2)

#>             ID           CCS_LVL_3_LABEL era firstCaseDate endCaseDate count
#> 1: 10003439853           Normal delivery   1    2008-04-23  2008-04-23     1
#> 2: 10003439853 Previous cesarean section   1    2008-04-22  2008-04-24     3
#> 3: 10003439853 Previous cesarean section   1    2008-04-22  2008-04-24     3
#> 4: 10006205602 Previous cesarean section   1    2008-05-27  2008-05-27     1
#> 5:   100085361           Normal delivery   1    2008-04-30  2008-04-30     1
#> 6:   100085361       Prolonged pregnancy   1    2008-04-29  2008-04-29     1
#>    period selectedCase
#> 1: 0 days         <NA>
#> 2: 1 days         <NA>
#> 3: 1 days         <NA>
#> 4: 0 days         <NA>
#> 5: 0 days         <NA>
#> 6: 0 days         <NA>