Plot of diagnostic categories

Histogram plot of diagnostic categories

plotDiagCat(groupedDataWide, idColName, groupColName = NULL, topN = 10,
  limitFreq = 0.01, pvalue = 0.05)

Arguments

groupedDataWide
GroupedData file from functions of code classification (four stPercentagegies)

limitFreq
The minimum frequency set to "0.01"; In other words, the limit at the same diagnostic category must have 1 percent patient in the total patient.

pvalue
p value of chisq.test

TopN
Default is Top "10"

Value

A histogram plot and a data.table for classifying data.

Details

An overview of grouping category of the diagnostic code summarizes the percentage of result in histogram plot. User can observe the number of diagnostic categories in their dataset through this function. This function also verifies by Chi-square test and Fisher’s exact test statistical significantly different diagnostic categories between case and control. The default level of significance considered at 5

See also

Examples

# sample file for example
head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

# Convert long format of grouped data into wide binary format

groupedDataWide <- groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                                         icd10usingDate = "2015-10-01",
                                         groupDataType = elix,
                                         isDescription = FALSE)

#> Error in groupedDataLongToWide(sampleDxFile, ID, ICD, Date, icd10usingDate = "2015-10-01",     groupDataType = elix, isDescription = FALSE): unused arguments (icd10usingDate = "2015-10-01", groupDataType = elix, isDescription = FALSE)

# plot of top 10 common grouped categories and a list of the detail of grouped categories

plot1 <- plot_groupedData(groupedDataWide = groupedDataWide,
                          TopN = 10,
                          limitFreq = 0.01)

#> Error in plot_groupedData(groupedDataWide = groupedDataWide, TopN = 10,     limitFreq = 0.01): could not find function "plot_groupedData"

plot1

#> Error in eval(expr, envir, enclos): object 'plot1' not found

# Select case with "Diseases of the urinary system" by level 2 of CCS classification

selectedCaseFile <- selectCases(dxDataFile = sampleDxFile,
                                idColName = ID,
                                icdColName = ICD,
                                dateColName = Date,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl2,
                                caseCondition = "Diseases of the urinary system",
                                caseCount = 1)

#> Wrong ICD format: total 9 ICD codes (the number of occurrences is in brackets)

#> c("A0.11 (20)", "E114 (8)", "Z9.90 (6)", "F42 (6)", "001 (5)", "75.52 (4)", "755.2 (3)", "123.45 (3)", "7552 (2)")

#> 

#> Wrong ICD version: total 7 ICD codes (the number of occurrences is in brackets)

#> c("V27.0 (18)", "A01.05 (8)", "42761 (7)", "V24.1 (6)", "A0105 (5)", "E03.0 (4)", "650 (4)")

#> 

#> Warning: The ICD mentioned above matches to "NA" due to the format or other issues.

#> Warning: "Wrong ICD format" means the ICD has wrong format

#> Warning: "Wrong ICD version" means the ICD classify to wrong ICD version (cause the "icd10usingDate" or other issues)

# Convert the long format of grouped data into a wide binary format with selected case

groupedDataWide <- groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                                         icd10usingDate = "2015-10-01",
                                         groupDataType = elix,
                                         isDescription = FALSE,
                                         selectedCaseFile = selectedCaseFile)

#> Error in groupedDataLongToWide(sampleDxFile, ID, ICD, Date, icd10usingDate = "2015-10-01",     groupDataType = elix, isDescription = FALSE, selectedCaseFile = selectedCaseFile): unused arguments (icd10usingDate = "2015-10-01", groupDataType = elix, isDescription = FALSE)

# plot of top 10 common grouped categories and a list of the detail of grouped categories

plot2 <- plot_groupedData(groupedDataWide = groupedDataWide,
                          TopN = 10,
                          limitFreq = 0.01,
                          pvalue = 0.05)

#> Error in plot_groupedData(groupedDataWide = groupedDataWide, TopN = 10,     limitFreq = 0.01, pvalue = 0.05): could not find function "plot_groupedData"

plot2

#> Error in eval(expr, envir, enclos): object 'plot2' not found