Code classification for CCS

The CCS classification for ICD-9 and ICD-10 codes is a procedure categorization scheme that can employ in many types of projects analyzing data on procedures

icdPrToCCS(prDataFile, idColName, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL, isDescription = TRUE)

icdPrToCCSLvl(prDataFile, idColName, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL, CCSLevel = 1,
  isDescription = TRUE)

Arguments

prDataFile
A file of clinical procedure data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A MemberID column of prDataFile

icdColName
A ICD column of prDataFile

dateColName
A Date column of prDataFile

icd10usingDate
ICD-10 using date

isDescription
Category or description of standard classification methods for ICD procedure codes. By default, it set to True (standard description).

CCSLevel
By default, it set to 1. CCS multiple level:1~3, CCS for ICD-10-CM only has 1~2 multiple levels.

Value

Two new data.tables. 1) groupedDT: Based on prDataFile with two new columns for uniform format procedure codes and classified standard categories. 2) Error: Potential error codes from PrUniform.

Details

These CCS functions collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD procedure codes.

See also

Examples

# sample file for example

head(samplePrFile)

#>    ID   ICD       Date
#> 1:  B  5681 2008-01-14
#> 2:  A  9774 2009-01-11
#> 3:  B 44.99 2009-05-10
#> 4:  C 07.59 2009-01-21
#> 5:  B  0205 2008-07-06
#> 6:  B  8812 2007-06-27

# Group procedure codes into single level of CCS classification

icdPrToCCS(samplePrFile, ID, ICD, Date, "2015-10-01", TRUE)

#> Error in `[.data.table`(prDataFile, , dataCol, with = FALSE): column(s) not found: "2015-10-01"

# Group procedure codes into multiple levels of CCS classification

icdPrToCCSLvl(samplePrFile, ID, ICD, Date, "2015-10-01", 2, TRUE)

#> Error in `[.data.table`(prDataFile, , dataCol, with = FALSE): column(s) not found: "2015-10-01"