Code classification for customized group

Researches can define the grouping categories and therefore have more flexible for grouping ICD diagnostic codes.

icdDxToCustom(dxDataFile, idColName, icdColName, dateColName,
  customGroupingTable)

icdDxToCustomGrep(dxDataFile, idColName, icdColName, dateColName,
  customGroupingTable)

Arguments

dxDataFile
A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"

idColName
A column for MemberID of dxDataFile.

icdColName
A column for ICD of dxDataFile

dateColName
A column for Date of dxDataFile

customGroupingTable
Define the grouping categories. `icdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories. "ICD" defines a list of disease-related ICD codes. `icdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories. "grepICD" defines disease-related ICD code character strings containing a regular expression.

Value

Two new data.tables. 1) groupedDT: Based on dxDataFile with two new columns for uniform format diagnostic codes and classified standard categories. 2) summarised_groupedDT: Summarized the dataset groupedDT and sorted by memberID.

Details

There are two functions for customized defined grouping method, the customized category grouping is based on precise (`icdDxToCustom`) and fuzzy (`icdDxToCustomGrep`) grouping method, respectively.

See also

Examples

# sample file for example

head(sampleDxFile)

#>     ID  ICD       Date
#> 1:  A2 Z992 2020-05-22
#> 2:  A5 Z992 2020-01-24
#> 3:  A8 Z992 2015-10-27
#> 4: A13 Z992 2020-04-26
#> 5: A13 Z992 2025-02-02
#> 6: A15 Z992 2023-05-12

# Group diagnostic codes into "Chronic kidney disease" with precise grouping method

groupingTable <- data.frame(Group = rep("Chronic kidney disease",6),
                            ICD = c("N181","5853","5854","5855","5856","5859"),
                            stringsAsFactors = FALSE)

icdDxToCustom(sampleDxFile, ID, ICD, Date, CustomGroupingTable = groupingTable)

#> Error in icdDxToCustom(sampleDxFile, ID, ICD, Date, CustomGroupingTable = groupingTable): unused argument (CustomGroupingTable = groupingTable)

# Group diagnostic codes into "Chronic kidney disease" with fuzzy grouping method

grepTable <- data.frame(Group = "Chronic kidney disease",
                        grepIcd = "^585|^N18",
                        stringsAsFactors = FALSE)

icdDxToCustomGrep(sampleDxFile,ID, ICD, Date,CustomGroupingTable = grepTable)

#> Error in icdDxToCustomGrep(sampleDxFile, ID, ICD, Date, CustomGroupingTable = grepTable): unused argument (CustomGroupingTable = grepTable)